backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
httpx==0.25.0
sqlite-utils==3.35.1
python-dotenv==1.1.0